+++ Scraps.Common/Constants/RegexPatterns.cs
@@ -1,6 +1,6 @@
 ﻿#region License
 /// Scraps - Scrap.TF Raffle Bot
-/// Copyright(C) 2021  Caprine Logic
+/// Copyright(C) 2022 Caprine Logic
 
 /// This program is free software: you can redistribute it and/or modify
 /// it under the terms of the GNU General Public License as published by
@@ -22,14 +22,13 @@ namespace Scraps.Common.Constants
 {
     public class RegexPatterns
     {
-        public static readonly Regex CsrfRegex = new(@"ScrapTF\.User\.Hash = ""([a-f0-9]{64})"";", RegexOptions.Compiled);
-        public static readonly Regex RaffleEntryRegex = new(@"ScrapTF\.Raffles\.RedirectToRaffle\('([A-Z0-9]{6,})'\)", RegexOptions.Compiled);
-        public static readonly Regex RaffleHashRegex = new(@"EnterRaffle\('[A-Z0-9]{6,}', '([a-f0-9]{64})'", RegexOptions.Compiled);
-        public static readonly Regex RaffleLimitRegex = new(@"total=""(\d+)"" data-max=""(\d+)", RegexOptions.Compiled);
-        public static readonly Regex RaffleWithdrawRegex = new(@"ScrapTF\.Raffles\.WithdrawRaffle\('([A-Z0-9]{6,})'\)", RegexOptions.Compiled);
-        public static readonly Regex WonRafflesAlertRegex = new(@"You've won \d raffles? that must be withdrawn", RegexOptions.Compiled);
-        public static readonly Regex HoneypotRaffleStyleRegex = new(@"<style>\.enter-raffle-btns \.btn{(.*)}<\/style>", RegexOptions.Compiled);
-        public static readonly Regex HoneypotRaffleMaxEntriesRegex = new(@"data-max=""(\d{1,})""", RegexOptions.Compiled);
-        public static readonly Regex HoneypotRaffleBannedUsersRegex = new(@"<img class='tiny-raffle-avatar\s?' style='border-color:\s?#CC1100;?' src='(.*)' loading=""lazy""\s?\/>", RegexOptions.Compiled);
+        public static readonly Regex CSRF = new(@"ScrapTF\.User\.Hash = ""([a-f0-9]{64})"";", RegexOptions.Compiled);
+        public static readonly Regex RAFFLE_ENTRY = new(@"ScrapTF\.Raffles\.RedirectToRaffle\('([A-Z0-9]{6,})'\)", RegexOptions.Compiled);
+        public static readonly Regex RAFFLE_HASH = new(@"EnterRaffle\('[A-Z0-9]{6,}', '([a-f0-9]{64})'", RegexOptions.Compiled);
+        public static readonly Regex RAFFLE_LIMIT = new(@"total=""(\d+)"" data-max=""(\d+)", RegexOptions.Compiled);
+        public static readonly Regex WON_RAFFLES_ALERT = new(@"You've won \d raffles? that must be withdrawn", RegexOptions.Compiled);
+        public static readonly Regex HONEYPOT_RAFFLE_STYLE = new(@"<style>\.enter-raffle-btns \.btn{(.*)}<\/style>", RegexOptions.Compiled);
+        public static readonly Regex HONEYPOT_RAFFLE_MAX_ENTRIES = new(@"data-max=""(\d{1,})""", RegexOptions.Compiled);
+        public static readonly Regex HONEYPOT_RAFFLE_BANNED_USERS = new(@"<img class='tiny-raffle-avatar\s?' style='border-color:\s?#CC1100;?' src='(.*)' loading=""lazy""\s?\/>", RegexOptions.Compiled);
     }
 }