+++ Scraps.Common/Constants/Xpaths.cs
@@ -1,6 +1,6 @@
 ﻿#region License
 /// Scraps - Scrap.TF Raffle Bot
-/// Copyright(C) 2021  Caprine Logic
+/// Copyright(C) 2022 Caprine Logic
 
 /// This program is free software: you can redistribute it and/or modify
 /// it under the terms of the GNU General Public License as published by
@@ -20,6 +20,6 @@ namespace Scraps.Common.Constants
 {
     public class Xpaths
     {
-        public const string UnenteredRaffles = "//*[contains(@class, 'panel-raffle') and not(contains(@class, 'raffle-entered'))]";
+        public const string UNENTERED_RAFFLES = "//*[contains(@class, 'panel-raffle') and not(contains(@class, 'raffle-entered'))]";
     }
 }