+++ Scraps.GUI/Bootstrapper.cs
@@ -1,6 +1,6 @@
 #region License
 /// Scraps - Scrap.TF Raffle Bot
-/// Copyright(C) 2021  Caprine Logic
+/// Copyright(C) 2022 Caprine Logic
 
 /// This program is free software: you can redistribute it and/or modify
 /// it under the terms of the GNU General Public License as published by
@@ -56,8 +56,8 @@ namespace Scraps.GUI
         {
             foreach (string path in new string[]
             {
-                Paths.LogsPath,
-                Paths.DataPath,
+                Paths.LOGS_PATH,
+                Paths.DATA_PATH,
             })
             {
                 if (!Directory.Exists(path))