+++ Scraps.GUI/Forms/SettingsForm.Designer.cs
@@ -276,6 +276,7 @@ namespace Scraps.GUI.Forms
             this.Name = "SettingsForm";
             this.StartPosition = System.Windows.Forms.FormStartPosition.CenterParent;
             this.Text = "Scraps Settings";
+            this.FormClosing += new System.Windows.Forms.FormClosingEventHandler(this.SettingsForm_FormClosing);
             this.groupBox2.ResumeLayout(false);
             this.groupBox2.PerformLayout();
             ((System.ComponentModel.ISupportInitialize)(this._JoinDelayInput)).EndInit();