+++ Scraps.GUI/Honeypot/Checks/BannedEntriesCheck.cs
@@ -1,6 +1,6 @@
 ﻿#region License
 /// Scraps - Scrap.TF Raffle Bot
-/// Copyright(C) 2021  Caprine Logic
+/// Copyright(C) 2022 Caprine Logic
 
 /// This program is free software: you can redistribute it and/or modify
 /// it under the terms of the GNU General Public License as published by
@@ -38,7 +38,7 @@ namespace Scraps.GUI.Honeypot.Checks
 
         public void Check()
         {
-            var entries = RegexPatterns.HoneypotRaffleBannedUsersRegex.Matches(_html);
+            var entries = RegexPatterns.HONEYPOT_RAFFLE_BANNED_USERS.Matches(_html);
 
             if (entries.Count > 1)
             {